experimental code for a graphical rtv menu
tried using sprites but those don't scale and scmapdb.spr will end up being massive if it supports multiple scales.

also tried attaching sprites to a weapon model but that's limited to 4 slots and you can't change the fov for some weapons or in observer mode, so the scaling problem is still there.