playing any map in a series counts as a play of that series
I always join half-life late and don't want it being nommed every day just because I wasn't there for the first map.

series_maps.py must be run to generate the series_maps.txt file, and you'll need to edit it because the detection isn't perfect and will combine things like "quarter" and "osprey" despite the trigger_changelevel in that map not working.